backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
langgraph==0.2.14
langchain==0.2.16
langchain-openai==0.1.23
pydantic==2.9.0
python-dotenv==1.0.0
httpx==0.27.0
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.1
pandas==2.2.2
feedparser==6.0.11
pymongo==4.6.3
pydantic[email]
